회원정보 보기시 권한 검증

+++ src/main/java/com/example/prj_practice_back/controller/MemberController.java
@@ -63,7 +63,14 @@ public class MemberController {
     }
 
     @GetMapping
-    public ResponseEntity<Member> view(String id) {
+    public ResponseEntity<Member> view(String id,
+                                       @SessionAttribute(value = "login", required = false) Member login) {
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        if (!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         // TODO : 로그인 했는 지? -> 안했으면 401
         // TODO : 자기 정보인지? -> 아니면 403
 
@@ -73,9 +80,18 @@ public class MemberController {
     }
 
     @DeleteMapping
-    public ResponseEntity delete(String id) {
+    public ResponseEntity delete(String id,
+                                 @SessionAttribute(value = "login",required = false)Member login) {
+
         // TODO : 로그인 했는 지? -> 안했으면 401
         // TODO : 자기 정보인지? -> 아니면 403
+
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        if(!service.deleteAccess(login, id)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
 
         if (service.deleteMember(id)) {
             return ResponseEntity.ok().build();
@@ -86,7 +102,14 @@ public class MemberController {
     }
 
     @PutMapping("edit")
-    public ResponseEntity edit(@RequestBody Member member) {
+    public ResponseEntity edit(@RequestBody Member member,
+                               @SessionAttribute(value = "login",required = false)Member login) {
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        if (!service.hasAccess(member.getId(),login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         // TODO: 로그인 했는지, 자기 정보인지
 
         if (service.update(member)) {
